src/requirements.txt
Django==2.2.4
django-debug-toolbar==2.0
django-dotenv==1.4.2
django-extensions==2.2.1
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3
Werkzeug==0.15.5
django-extra-views==0.12.0
celery==4.3.0
redis==3.3.8
tablib-0.13.0
python-magic==0.4.15